% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatcher.R
\name{analyzePictureCol}
\alias{analyzePictureCol}
\title{Analyze picture colors}
\usage{
analyzePictureCol(
  file_path = NULL,
  link = NULL,
  m = NULL,
  reference_space = NULL,
  keep_extremes = TRUE
)
}
\arguments{
\item{file_path}{a character indicating a path to a .JPG or .PNG file. Either "file"
or "link" must be specified.}

\item{link}{a character indicating a URL to a .JPG or .PNG file. Either "link"
or "file" must be specified.}

\item{m}{a raster object created by either \code{readJPEG} or \code{readPNG}}

\item{reference_space}{a reference space array created by a call to \code{makeReferenceSpace()}.
Default is NULL, which internally loads a reference space comprising 2000
colors in DIN99 space.}

\item{keep_extremes}{logical, should the extreme points of the space
(corresponding to any combination of full R, G, or B channel) be kept?
Default is TRUE.}
}
\value{
a named numeric vector of summary color quantification (number of
    pixels with that summary color), whose names are the hexadecimal values of
    colors.
}
\description{
Quantifies summary colors in a picture
}
\details{
This function decomposes a .JPG or .PNG file into component RGB values,
    assigns each of these values to a summary color as indicated in the reference
    space, and quantifies the number of pixels for each summary color. The result
    is unsorted.
}
\author{
Giuseppe D'Agostino
}
