% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_ops.R
\name{CIELabtoDIN99mod}
\alias{CIELabtoDIN99mod}
\title{CIELab to DIN99 transformation}
\usage{
CIELabtoDIN99mod(L, a, b)
}
\arguments{
\item{L}{numeric vector of L values}

\item{a}{numeric vector of a values}

\item{b}{numeric vector of b values}
}
\value{
a matrix with the same number of rows as the length of the vectors in
    DIN99 coordinates.
}
\description{
Transform L, a, b, to L99o, a99o, b99o
}
\details{
Internal use only. Slightly modified to take vectors and not single
    values as input. Originally present in the \code{colorscience} package
    by Jose Gama.
}
\references{
CIELAB to DIN99 coordinates, 2014 http://de.wikipedia.org/w/index.php?title=Diskussion:DIN99-Farbraum
}
\author{
Jose Gama, modified by Giuseppe D'Agostino
}
