% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatcher.R
\name{clusterKDIN}
\alias{clusterKDIN}
\title{K clusters in DIN99 space}
\usage{
clusterKDIN(sorted, n, sub, method)
}
\arguments{
\item{sorted}{picture analysis from \code{analyzePictureCol}}

\item{n}{a numeric, the final number of clusters}

\item{sub}{a numeric, most abundant colors per cluster}

\item{method}{a character, one of "kmeans" or "HC"}
}
\value{
a palette with most abundant and separated colors
}
\description{
Clusters a color palette in DIN99 color space returning k clusters
}
\details{
Internal use only.
}
\author{
Giuseppe D'Agostino
}
