% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatcher.R
\name{deltaE2000mod}
\alias{deltaE2000mod}
\title{DeltaE200 color difference}
\usage{
deltaE2000mod(Lab1, Lab2)
}
\arguments{
\item{Lab1}{matrix with L, a, b coordinates for one set of colors}

\item{Lab2}{matrix with L, a, b coordinates for another set of colors}
}
\value{
a vector of DeltaE2000 color distances
}
\description{
Calculates the Delta E (CIE 2000) color difference between two matrices in CIELab space
}
\details{
Internal use only. Slightly modified for vectorisation.
    Originally present in the \code{colorscience} package by Jose Gama.
}
\references{
Bruce Justin Lindbloom, 2013 Color Calculator http://www.brucelindbloom.com
}
\author{
Jose Gama, modified by Giuseppe D'Agostino
}
