% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatcher.R
\name{DIN99toCIELabmod}
\alias{DIN99toCIELabmod}
\title{DIN99 to CIELab transformation}
\usage{
DIN99toCIELabmod(L99o, a99o, b99o)
}
\arguments{
\item{L99o}{numeric vector of L99o values}

\item{a99o}{numeric vector of a99o values}

\item{b99o}{numeric vector of b99o values}
}
\value{
a matrix with the same number of rows as the length of the vectors in
    LAB coordinates.
}
\description{
Transform L99o, a99o, b99o to L, a, b
}
\details{
Internal use only. Slightly modified to take vectors and not single
    values as input. Originally present in the \code{colorscience} package
    by Jose Gama.
}
\references{
DIN99 to CIELAB coordinates, 2014 http://de.wikipedia.org/w/index.php?title=Diskussion:DIN99-Farbraum
}
\author{
Jose Gama, modified by Giuseppe D'Agostino
}
