% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatcher.R
\name{filterColors}
\alias{filterColors}
\title{Filter colors}
\usage{
filterColors(colors, on, type, filter_sd)
}
\arguments{
\item{colors}{hex code vector of colors}

\item{on}{character, one of "L" (luminance) or "C" (chroma)}

\item{type}{character, one of "dark" or "bright" for "L",
"high" or "low" for "C", and "both" for either value of \code{on}}

\item{filter_sd}{numeric, the number of SD/MAD for filtering}
}
\value{
a vector of indices for filtering colors
}
\description{
Filters colors based on Luminace or Chroma
}
\details{
internal use only.
}
\author{
Giuseppe D'Agostino
}
