% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatcher.R
\name{getContinuousPalette}
\alias{getContinuousPalette}
\title{Get continuous palettes}
\usage{
getContinuousPalette(pal, n = 10, tries = 1, plot = TRUE)
}
\arguments{
\item{pal}{a hex code vector with the palette}

\item{n}{a numeric, the final number of palette colors. Default is 10}

\item{tries}{numeric, how many times should the algorithm run?}

\item{plot}{logical, should the resulting palettes be plotted?}
}
\value{
a nested list of length \code{tries}, each element containing:
\itemize{
   \item{the original palette in DIN99 space (\code{original_space})}
   \item{the palette space after clustering (\code{palette_space})}
   \item{the principal curve (\code{principal_curve})}
   \item{the curve colors (\code{principal_curve_rgb})}
   \item{the final palette (\code{palette})}
   }
}
\description{
Generates continuous palettes from a categorical palette
}
\details{
todo
}
