% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatcher.R
\name{getKNN}
\alias{getKNN}
\title{Get kNN graph}
\usage{
getKNN(coords, k = 2)
}
\arguments{
\item{coords}{matrix with point coordinates}

\item{k}{number of nearest neighbors}
}
\value{
an \code{igraph} graph object with the k-NN graph
}
\description{
Builds a k-nearest neighbor graph from a set of points
}
\details{
Internal use only. Nothing fancy, just a brute force
    barebones implementation of k-NN.
}
