% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatcher.R
\name{getPalette}
\alias{getPalette}
\title{Get a summary palette}
\usage{
getPalette(
  analysis,
  n = 10,
  sub = 1,
  filter_luminance = "both",
  filter_chroma = NULL,
  filter_sd = 1.5,
  order = "L"
)
}
\arguments{
\item{analysis}{a named numeric with the results of a color analysis, i.e. the output
of \code{analyzePaintingCol}}

\item{n}{a numeric, the number of major colors (k-means clusters)}

\item{sub}{a numeric, the number o minor colors (top abundant colors per cluster)}

\item{filter_luminance}{character, one of "dark", "bright", or "both". Removes
colors within a certain number of standard deviations from the mean of the
luminance values. "dark" keeps darker colors, "bright" keeps brighter colors,
"both" applies both filters. Default is "both".}

\item{filter_chroma}{character, one of "low", "high", or "both". Removes
colors within a certain number of median absolute deviations from the median
of the chroma values. "low" keeps less saturated colors, "high" keeps
more saturated colors, "both" applies both filters. Default is NULL.}

\item{filter_sd}{numeric, the number of standard deviations and/or median absolute
deviations above/below which filters are applied. Default is 1.5.}

\item{order}{character, the dimension of HCL space along which the final palette
should be ordered. One of "H" (hue), "C" (chroma), "L" (luminance). The
final order has only aesthetic purposes and does not change the palette.
Default is "L".}
}
\value{
a character vector of hexadecimal color values of length \code{n} * \code{sub}.
}
\description{
Automatically finds a color palette from a color analysis
}
\details{
This function performs several operations on a color analysis result
   to isolate a pre-determined number of the most representative colors in the
   analysis. First, the analysis results are sorted and the top 500 colors are retained;
   then they are optionally transformed to polarLUV coordinates (i.e. HCL) and
   filtered based on the distribution of luminance and/or chroma. Then, they
   are transformed to DIN99 space,and clustered using k-means clustering, using
   the user-defined parameter \code{n} as number of clusters. Within each
   cluster, the top \code{sub} colors (as defined by the analysis) are kept.
   The palette is then optionally ordered according to one of hue, chroma, or
   luminance.
}
\author{
Giuseppe D'Agostino
}
