% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatcher.R
\name{hClusterPalette}
\alias{hClusterPalette}
\title{Palette hierarchical clustering}
\usage{
hClusterPalette(palette, n)
}
\arguments{
\item{palette}{a vector of colors as hexadecimal values}

\item{n}{a numeric, desired number of clusters}
}
\value{
a \code{data.frame} with colors and clusters.
}
\description{
Applies hierarchical clustering to a palette returning clusters
}
\details{
Internal use only.
}
\author{
Giuseppe D'Agostino
}
