% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatcher.R
\name{makeReferenceSpace}
\alias{makeReferenceSpace}
\title{Make reference space}
\usage{
makeReferenceSpace(k = 1000, space = "LAB", verbose = FALSE)
}
\arguments{
\item{k}{a numeric indicating the final number of summary colors. Default is 1000}

\item{space}{a character, one of "LAB" or "DIN99"}

\item{verbose}{logical, should the function output information about the progress? Default is FALSE}
}
\value{
a cube 3D \code{array} of side 256 in which every value corresponds to a
    summarized color for that RGB coordinate. Values are encoded as native integers.
}
\description{
Build a summarized color reference space
}
\details{
This function creates the reference space to assign single RGB colors
    (defined by a triplet of R, G, B coordinates between 0 and 255) to one of
    a number (k) of clusters. Clusters are defined via k-means within one of
    two perceptual color spaces, CIELab or DIN99. The function builds the RGB
    cube first, then transforms it to either CIELab or DIN99 coordinates, performs
    k-means clustering and takes cluster centroid coordinates as summary colors.
    Summary colors are then assigned back to the original RGB cube. For a particular
    RGB triplet the corresponding position in the array is matched with the summary
    color (centroid of the clustering it was included in).
}
\note{
Since the number \code{k} of clusters is usually high (1000-3000) and
    distances are regularly spaced, the Hartigan-Wong algorithm may not converge
    in Quick-Transfer stage. This will throw a warning, but it does not seem to
    affect the results for this application.
}
\author{
Giuseppe D'Agostino
}
