% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatcher.R
\name{getPalette}
\alias{getPalette}
\title{Get a summary palette}
\usage{
getPalette(
  file_path = NULL,
  link = NULL,
  m = NULL,
  analysis = NULL,
  reference_space = NULL,
  method = "HC",
  n = 10,
  sub = 1,
  ntop = 2000,
  filter_luminance = "both",
  filter_chroma = NULL,
  filter_sd = 1.5,
  order = "L",
  keep_extremes = TRUE,
  optimize = FALSE,
  DE2000_target = 4
)
}
\arguments{
\item{file_path}{a character indicating a path to a .JPG or .PNG file.
Either \code{file_path},  \code{link} or  \code{m} must be specified.
Default is NULL.}

\item{link}{a character indicating a URL to a .JPG or .PNG file. Either
\code{file_path},  \code{link} or  \code{m} must be specified.
Default is NULL.}

\item{m}{a raster object created by either \code{readJPEG} or \code{readPNG}.
Either \code{file_path},  \code{link} or  \code{m} must be specified.
Default is NULL.}

\item{analysis}{a named numeric with the results of a color analysis, i.e. the output
of \code{analyzePaintingCol}. Default is NULL.}

\item{reference_space}{a reference space array created by a call to \code{makeReferenceSpace()}.
Default is NULL, which internally loads a reference space comprising 4096
colors in DIN99 space.}

\item{method}{a character specifying the method for clustering:
"HC" for hierarchical clustering, "kmeans" for k-means clustering, and
"kmeans_classic" for a legacy implementation. Default is "HC".}

\item{n}{a numeric, the number of major colors (k-means clusters)}

\item{sub}{a numeric, the number o minor colors (top abundant colors per cluster)}

\item{ntop}{a numeric, the number of top colors to use for the initial clustering.
Default is 2000.}

\item{filter_luminance}{character, one of "dark", "bright", or "both". Removes
colors within a certain number of standard deviations from the mean of the
luminance values. "dark" keeps darker colors, "bright" keeps brighter colors,
"both" applies both filters. Default is "both".}

\item{filter_chroma}{character, one of "low", "high", or "both". Removes
colors within a certain number of median absolute deviations from the median
of the chroma values. "low" keeps less saturated colors, "high" keeps
more saturated colors, "both" applies both filters. Default is NULL.}

\item{filter_sd}{numeric, the number of standard deviations and/or median absolute
deviations above/below which filters are applied. Default is 1.5.}

\item{order}{character, the dimension of HCL space along which the final palette
should be ordered. One of "H" (hue), "C" (chroma), "L" (luminance). The
final order has only aesthetic purposes and does not change the palette.
Default is "L".}

\item{keep_extremes}{logical, passed to \code{analyzePictureCol}}

\item{optimize}{logical, should the palette be optimized for a minimum
color difference (specified by \code{DE2000_target})? Default is FALSE.}

\item{DE2000_target}{the minimum DeltaE2000 difference for optimization.}
}
\value{
a character vector of hexadecimal color values of length \code{n} * \code{sub}.
}
\description{
Automatically finds a color swatch from a color analysis
}
\details{
This function performs several operations on a color analysis result
   to isolate a pre-determined number of the most representative colors in the
   analysis. First, the analysis results are sorted and the \code{ntop} top colors
   are retained; then they are optionally transformed to polarLUV coordinates
   (i.e. HCL) and filtered based on the distribution of luminance and/or chroma.
   Then, they are transformed to DIN99 space,and clustered using either
   hierarchical or k-means clustering, using the user-defined parameter \code{n}
   as number of clusters. Another method that does not use the reference space
   is "kmeans_classic", which applies k-means directly to the image (after
   the RGB values have been transformed to DIN99 space).
   Within each cluster, the top \code{sub} colors (as defined by the analysis)
   are kept. Optionally, the palette can be optimized by running the same procedure
   iteratively until a minimum DeltaE2000 color difference is reached. Optimization
   is stopped after 100 unsuccessful attempts, at which point it is advisable to
   lower the minimum difference. It is important to note that the bigger the
   palette, the lowest the minimum difference between colors will be.
   The palette is then optionally ordered according to one of hue, chroma, or
   luminance.
}
\author{
Giuseppe D'Agostino
}
