% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatcher.R
\name{getPaletteDistances}
\alias{getPaletteDistances}
\title{Get palette distances}
\usage{
getPaletteDistances(palette)
}
\arguments{
\item{palette}{a vector of colors as hexadecimal values}
}
\value{
a \code{data.frame} with color pairs and their Delta E (DIN99) color
   difference, i.e. the Euclidean distance in DIN99 color space.
}
\description{
Calculate pairwise differences between colors in a palette
}
\details{
Internal use only.
}
\author{
Giuseppe D'Agostino
}
