% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatcher.R
\name{plotWithPal}
\alias{plotWithPal}
\title{Plot a picture with its palette}
\usage{
plotWithPal(
  file_path = NULL,
  link = NULL,
  m = NULL,
  reference_space = NULL,
  method = "HC",
  n = 20,
  sub = 1,
  filter_luminance = "both",
  filter_chroma = NULL,
  filter_sd = 1.5,
  order = "L",
  optimize = FALSE,
  DE_target = 4,
  bg_color = "white",
  keep_extremes = TRUE,
  title = NULL
)
}
\arguments{
\item{file_path}{a character indicating a path to a .JPG or .PNG file. Either "file"
or "link" must be specified.}

\item{link}{a character indicating a URL to a .JPG or .PNG file. Either "link"
or "file" must be specified.}

\item{m}{a raster object created by either \code{readJPEG} or \code{readPNG}}

\item{reference_space}{a reference space array created by a call to \code{makeReferenceSpace()}}

\item{method}{a character specifying the method for clustering:
"HC" for hierarchical clustering, "kmeans" for k-means clustering, and
"kmeans_classic" for a legacy implementation. Default is "HC".}

\item{n}{a numeric, the number of major colors (clusters)}

\item{sub}{a numeric, the number o minor colors (top abundant colors per cluster)}

\item{filter_luminance}{character, one of "dark", "bright", or "both". Removes
colors within a certain number of standard deviations from the mean of the
luminance values. "dark" keeps darker colors, "bright" keeps brighter colors,
"both" applies both filters. Default is "both".}

\item{filter_chroma}{character, one of "low", "high", or "both". Removes
colors within a certain number of median absolute deviations from the median
of the chroma values. "low" keeps less saturated colors, "high" keeps
more saturated colors, "both" applies both filters. Default is NULL.}

\item{filter_sd}{numeric, the number of standard deviations and/or median absolute
deviations above/below which filters are applied. Default is 1.5.}

\item{order}{character, the dimension of HCL space along which the final palette
should be ordered. One of "H" (hue), "C" (chroma), "L" (luminance). The
final order has only aesthetic purposes and does not change the palette.
Default is "L".}

\item{optimize}{logical, should the palette be optimized for maximal difference?
Default is FALSE}

\item{DE_target}{numeric, the target minimum color difference for optimization.
Default is 4. Only relevant if \code{optimize} is TRUE.}

\item{bg_color}{the background color (as named color, hexadecimal or rgb).
Default is "white".}

\item{keep_extremes}{logical, passed to \code{analyzePictureCol} if needed.}

\item{title}{character, the title of the plot.}
}
\value{
a \code{ggplot2} plot with the picture and the corresponding palette.
}
\description{
Plots a picture and the resulting summary palette side by side
}
\details{
This function is a wrapper around ]code{analyzePaintingCol} and
    \code{getPalette} that allows the user to select a picture, extract the
    palette and plot them side by side.
}
\author{
Giuseppe D'Agostino
}
